LayoutTests/platform/ios-sim/storage/indexeddb/intversion-encoding-expected.txt

CONSOLE MESSAGE: line 19: TypeError: undefined is not an object (evaluating 'request.onblocked = unexpectedBlockedCallback')
Test that integer versions are encoded/decoded consistently.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "intversion-encoding.html"
indexedDB.deleteDatabase(dbname)
FAIL indexedDB.deleteDatabase(dbname) threw exception TypeError: null is not an object (evaluating 'indexedDB.deleteDatabase')
FAIL Unexpected error: TypeError: undefined is not an object (evaluating 'request.onblocked = unexpectedBlockedCallback')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
